Debugging

+++ src/misc/nugget.lua
@@ -9,34 +9,27 @@ while true do
     end
     
     for i = 1, chest.getInventorySize() do
-        local info = chest.getStackInSlot(i)
-        print(i)
---        for k, v in pairs(info) do
---        print(k .. ":" .. v)
---        end
-         if info then
-            if (string.find(info["name"], "Nugget")) then
-                if (info["qty"] > 9) then
-                	turtle.select(1)
-                	print("push item into slot")
-                    turtle.suck()
-                    local no = turtle.getItemCount(1)
-                    for j, k in ipairs(craft_slots) do
-                        turtle.transferTo(k, no/9)
-                    end
-                    turtle.craft()
-                    -- Drop the ingots into the chest
-                    turtle.select(2)
-                    turtle.turnRight()
-                    turtle.drop()
-                    turtle.turnLeft()
-                    -- Drop the leftover nuggets back into the chest
-                    turtle.select(1)
-                    turtle.drop()
-                end
-            end
-        end
-        chest.condenseItems()
+		local info = chest.getStackInSlot(i)
+		print("Slot number is now: " .. i)
+		if (info ~= nil and (string.find(info["name"], "Nugget")) and info["qty"] > 9) then
+			turtle.select(1)
+			print("push item into slot")
+			turtle.suck()
+			local no = turtle.getItemCount(1)
+			for j, k in ipairs(craft_slots) do
+			    turtle.transferTo(k, no/9)
+			end
+			turtle.craft()
+			-- Drop the ingots into the chest
+			turtle.select(2)
+			turtle.turnRight()
+			turtle.drop()
+			turtle.turnLeft()
+			-- Drop the leftover nuggets back into the chest
+		    turtle.select(1)
+		    turtle.drop()
+		end
+	    chest.condenseItems()
     end
     sleep(1200)
 end
